requirements.txt
Flask
Flask-Session
requests
gunicorn
python-dotenv
psycopg2